refactor(memory): split memory kernel merge-gate modules [impact-checked] [structural-delete-approved]
