+++ dharma_swarm/context_compiler.py
@@ -5,13 +5,15 @@ from __future__ import annotations
 import hashlib
 import json
 import logging
-import os
 from dataclasses import dataclass, field
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Callable, Iterable
 
 from dharma_swarm.memory_lattice import MemoryLattice, MemoryRecallHit
+from dharma_swarm.memory_kernel.context_compiler_shadow import (
+    run_memory_kernel_context_compiler_shadow,
+)
 from dharma_swarm.provider_policy import ProviderPolicyRouter, ProviderRouteRequest
 from dharma_swarm.runtime_state import (
     ArtifactRecord,
@@ -82,23 +84,6 @@ def _context_scan_metadata(rendered_text: str) -> dict[str, Any]:
         "scanner": "dharma_swarm.injection_scanner.scan_content",
     }
 
-
-def _truthy_env(name: str) -> bool:
-    return os.environ.get(name, "").strip().lower() in {"1", "true", "yes", "on"}
-
-
-def _csv_env(name: str) -> tuple[str, ...]:
-    value = os.environ.get(name, "")
-    return tuple(item.strip() for item in value.split(",") if item.strip())
-
-
-def _path_env(name: str) -> Path | None:
-    value = os.environ.get(name, "").strip()
-    return Path(value).expanduser() if value else None
-
-
-def _shadow_atom_budget(token_budget: int) -> int:
-    return max(1, min(50, max(1, int(token_budget)) // 100))
 
 
 @dataclass(frozen=True)
@@ -258,7 +243,7 @@ class ContextCompiler:
                 )
                 await self.runtime_state.init_db()
                 saved = await self.runtime_state.record_context_bundle(bundle)
-                await self._run_memory_kernel_shadow(
+                await run_memory_kernel_context_compiler_shadow(
                     saved,
                     query=query or task_description,
                     token_budget=token_budget,
@@ -272,7 +257,7 @@ class ContextCompiler:
         # Return frozen snapshot if available (preserves prompt cache)
         if not force_refresh and session_id in self._frozen_bundles:
             frozen = self._frozen_bundles[session_id]
-            await self._run_memory_kernel_shadow(
+            await run_memory_kernel_context_compiler_shadow(
                 frozen,
                 query=self._compose_recall_query(
                     operator_intent=operator_intent,
@@ -428,7 +413,7 @@ class ContextCompiler:
                     updated_at=created_at,
                 )
             )
-        await self._run_memory_kernel_shadow(
+        await run_memory_kernel_context_compiler_shadow(
             saved,
             query=recall_query,
             token_budget=token_budget,
@@ -438,77 +423,6 @@ class ContextCompiler:
             callback=memory_kernel_shadow_callback,
         )
         return saved
-
-    async def _run_memory_kernel_shadow(
-        self,
-        bundle: ContextBundleRecord,
-        *,
-        query: str | None,
-        token_budget: int,
-        enabled: bool,
-        home: Path | None,
-        surface_ids: tuple[str, ...],
-        callback: Callable[[object], object] | None,
-    ) -> None:
-        """Run MemoryKernel context parity without altering compiled context."""
-
-        env_enabled = _truthy_env("DHARMA_MEMORY_KERNEL_CONTEXT_COMPILER_SHADOW")
-        global_env_enabled = _truthy_env("DHARMA_MEMORY_KERNEL_CONTEXT_SHADOW")
-        if not enabled and not env_enabled and not global_env_enabled:
-            return
-        try:
-            from dharma_swarm.memory_kernel import (
-                CensusConfig,
-                MemoryContextBudget,
-                MemoryKernel,
-                MemoryKernelConfig,
-                run_context_parity_eval,
-            )
-
-            resolved_surfaces = (
-                surface_ids
-                or _csv_env("DHARMA_MEMORY_KERNEL_CONTEXT_COMPILER_SURFACES")
-                or _csv_env("DHARMA_MEMORY_KERNEL_CONTEXT_SURFACES")
-            )
-            resolved_home = home or _path_env("DHARMA_MEMORY_KERNEL_HOME")
-            kernel = None
-            if resolved_home is not None and resolved_surfaces:
-                kernel = MemoryKernel(
-                    MemoryKernelConfig(
-                        census=CensusConfig(
-                            repo_root=Path.cwd(),
-                            home=resolved_home,
-                            include_discovered=False,
-                        )
-                    )
-                )
-            atom_budget = _shadow_atom_budget(token_budget)
-            report = run_context_parity_eval(
-                query=query,
-                current_context_text=bundle.rendered_text,
-                state_dir=None,
-                memory_kernel=kernel,
-                memory_surface_ids=resolved_surfaces,
-                budget=MemoryContextBudget(
-                    max_candidate_atoms=atom_budget,
-                    max_admitted_atoms=max(1, min(atom_budget, 8)),
-                    include_content=False,
-                ),
-                allow_current_semantic_search=False,
-            )
-            if callback:
-                callback_result = callback(report)
-                if hasattr(callback_result, "__await__"):
-                    await callback_result
-            logger.debug(
-                "MemoryKernel context compiler shadow parity: bundle_id=%s hard_failures=%s warnings=%s metrics=%s",
-                bundle.bundle_id,
-                report.hard_failure_count,
-                report.warning_count,
-                report.metric_count,
-            )
-        except Exception:
-            logger.debug("MemoryKernel context compiler shadow failed", exc_info=True)
 
     def _compose_recall_query(
         self,
