+++ dharma_swarm/memory_kernel/surface_specs_core.py
@@ -1,0 +1,682 @@
+"""Grouped MemoryKernel surface registry entries."""
+
+from __future__ import annotations
+
+from dharma_swarm.memory_kernel.atoms import (
+    AdapterMode,
+    AuthorityLevel,
+    MemorySurfaceRole,
+    MigrationStatus,
+    RiskLevel,
+    SurfaceCategory,
+    SurfaceStatus,
+    WriteMode,
+)
+from dharma_swarm.memory_kernel.surfaces import SurfaceSpec
+
+
+def core_surface_specs() -> tuple[SurfaceSpec, ...]:
+    return (
+        SurfaceSpec(
+            "home.dharma_root",
+            "{home}/.dharma",
+            "external",
+            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.UNKNOWN,
+            AdapterMode.METADATA_ONLY,
+            feeds=("surface_health",),
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Root container; register children through adapters.",
+        ),
+        SurfaceSpec(
+            "home.memory_plane",
+            "{home}/.dharma/db/memory_plane.db",
+            "dharma_swarm.engine.event_memory",
+            MemorySurfaceRole.MEMORY_PLANE,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.MEDIUM,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.READ_ONLY,
+            provenance_quality="medium",
+            known_writers=(
+                "engine/event_memory.py",
+                "engine/conversation_memory.py",
+                "engine/retrieval_feedback.py",
+                "memory_lattice.py",
+            ),
+            known_readers=("context.py", "memory_lattice.py", "memory_palace.py"),
+            feeds=("episodes", "source_chunks", "retrieval_feedback", "idea_shards"),
+            source_of_truth_for=("raw_events", "conversation_turns", "idea_shards"),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            sqlite_count_tables=(
+                "event_log",
+                "conversation_turns",
+                "idea_shards",
+                "idea_links",
+                "idea_uptake",
+                "retrieval_log",
+                "source_documents",
+                "source_chunks",
+            ),
+        ),
+        SurfaceSpec(
+            "home.runtime_state",
+            "{home}/.dharma/state/runtime.db",
+            "dharma_swarm.runtime_state",
+            MemorySurfaceRole.RUNTIME_STATE,
+            SurfaceCategory.RUNTIME_CONTROL,
+            AuthorityLevel.HIGH,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            provenance_quality="medium",
+            known_writers=("runtime_state.py", "telemetry_plane.py", "route_witness.py"),
+            known_readers=("context_compiler.py", "terminal_bridge_context.py"),
+            feeds=("runtime_events", "memory_facts", "memory_edges", "context_bundles"),
+            source_of_truth_for=("runtime_sessions", "routing_decisions", "context_bundles"),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            sqlite_count_tables=(
+                "session_events",
+                "sessions",
+                "routing_decisions",
+                "context_bundles",
+                "memory_facts",
+                "memory_edges",
+            ),
+        ),
+        SurfaceSpec(
+            "home.memory_db",
+            "{home}/.dharma/db/memory.db",
+            "dharma_swarm.memory",
+            MemorySurfaceRole.SESSION_MEMORY,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            known_writers=("memory.py", "swarm.py", "pulse.py"),
+            known_readers=("context.py", "swarm.py", "mcp_server.py"),
+            feeds=("episodes",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.LEGACY_TOLERATED,
+            sqlite_count_tables=("memories",),
+        ),
+        SurfaceSpec(
+            "home.bridges",
+            "{home}/.dharma/db/bridges.db",
+            "dharma_swarm.bridge_registry",
+            MemorySurfaceRole.KNOWLEDGE_GRAPH,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            provenance_quality="medium",
+            known_writers=("bridge_registry.py", "pruner.py"),
+            feeds=("cross_graph_bridge_edges",),
+            projection_of=("semantic_graphs", "runtime_facts", "ontology", "telos"),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            do_not_migrate_reason="Bridge edges connect graphs but are not canon by themselves.",
+            sqlite_count_tables=("bridge_edges",),
+        ),
+        SurfaceSpec(
+            "home.ecosystem_index",
+            "{home}/.dharma/db/ecosystem_index.db",
+            "dharma_swarm.ecosystem_index",
+            MemorySurfaceRole.PROJECTION_INDEX,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.NONE,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.METADATA_ONLY,
+            known_writers=("ecosystem_index.py",),
+            feeds=("ecosystem_search_pointers",),
+            projection_of=("workspace_files", "external_docs"),
+            canon_risk=RiskLevel.CRITICAL,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Full-text index over files; source files remain authority.",
+            sqlite_count_tables=("doc_meta",),
+        ),
+        SurfaceSpec(
+            "home.router_audit_log",
+            "{home}/.dharma/logs/router/routing_decisions.jsonl",
+            "dharma_swarm.route_witness",
+            MemorySurfaceRole.WITNESS,
+            SurfaceCategory.PROVENANCE,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            provenance_quality="medium",
+            known_writers=("route_witness.py",),
+            known_readers=("routing dashboards", "post-hoc audits"),
+            feeds=("routing_witness_events", "provider_attempts"),
+            source_of_truth_for=("router_audit_jsonl",),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.routing_memory",
+            "{home}/.dharma/logs/router/routing_memory.sqlite3",
+            "dharma_swarm.routing_memory",
+            MemorySurfaceRole.RUNTIME_STATE,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            provenance_quality="medium",
+            known_writers=("routing_memory.py",),
+            known_readers=("smart_router.py", "agent_runner.py"),
+            feeds=("provider_learning", "routing_stats", "routing_events"),
+            source_of_truth_for=("provider_lane_learning",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            sqlite_count_tables=("routing_stats", "routing_events"),
+        ),
+        SurfaceSpec(
+            "home.hibernation",
+            "{home}/.dharma/state/hibernation.db",
+            "dharma_swarm.hibernation",
+            MemorySurfaceRole.RUNTIME_STATE,
+            SurfaceCategory.RUNTIME_CONTROL,
+            AuthorityLevel.MEDIUM,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            provenance_quality="medium",
+            known_writers=("hibernation.py",),
+            feeds=("hibernation_jobs", "wake_conditions"),
+            source_of_truth_for=("hibernateable_job_state",),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            sqlite_count_tables=("hibernation_jobs",),
+        ),
+        SurfaceSpec(
+            "home.algedonic_signals",
+            "{home}/.dharma/algedonic_signals.jsonl",
+            "dharma_swarm.orchestrator",
+            MemorySurfaceRole.WITNESS,
+            SurfaceCategory.PROVENANCE,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            provenance_quality="medium",
+            known_writers=("orchestrator.py",),
+            feeds=("algedonic_events",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.smriti",
+            "{home}/.smriti/smriti.db",
+            "smriti",
+            MemorySurfaceRole.SESSION_MEMORY,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            provenance_quality="weak",
+            known_writers=("~/.claude/hooks/smriti_session_end.py",),
+            known_readers=("~/.claude/hooks/smriti_session_start.py",),
+            feeds=("episodes", "procedural_candidates"),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            sqlite_count_tables=("memories", "entities", "relations"),
+        ),
+        SurfaceSpec(
+            "home.conversation_log",
+            "{home}/.dharma/conversation_log",
+            "dharma_swarm.conversation_log",
+            MemorySurfaceRole.RAW_EPISODIC_LOG,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            known_writers=("conversation_log.py", "~/.claude/hooks/conversation_logger.py"),
+            known_readers=("api/routers/dashboard_new.py",),
+            feeds=("episodes",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            do_not_migrate_reason="Use streaming adapter; never eager-load 24 GB logs.",
+        ),
+        SurfaceSpec(
+            "home.knowledge_root",
+            "{home}/.dharma/knowledge",
+            "chetana",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.METADATA_ONLY,
+            provenance_quality="medium",
+            feeds=("knowledge_lifecycle_atoms",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            do_not_migrate_reason="Chetana is metabolism; staging is not canon.",
+        ),
+        SurfaceSpec(
+            "home.knowledge_staging",
+            "{home}/.dharma/knowledge/staging",
+            "chetana",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.STREAMING,
+            feeds=("staged_knowledge_atoms",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            do_not_migrate_reason="Staging is capture backlog and must not be treated as canon.",
+        ),
+        SurfaceSpec(
+            "home.knowledge_wiki",
+            "{home}/.dharma/knowledge/wiki",
+            "chetana",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.MEDIUM,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            feeds=("curated_knowledge_atoms",),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.knowledge_quarantine",
+            "{home}/.dharma/knowledge/quarantine",
+            "chetana",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.NONE,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            feeds=("rejected_or_quarantined_atoms",),
+            canon_risk=RiskLevel.CRITICAL,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Quarantine is explicitly non-authoritative.",
+        ),
+        SurfaceSpec(
+            "home.knowledge_extracts",
+            "{home}/.dharma/knowledge/extracts",
+            "chetana",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.STREAMING,
+            feeds=("evidence_candidates",),
+            projection_of=("conversation_logs", "source_material"),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Extracts are derived evidence; stream pointers only.",
+        ),
+        SurfaceSpec(
+            "home.knowledge_store",
+            "{home}/.dharma/db/knowledge_store.db",
+            "dharma_swarm.engine.knowledge_store",
+            MemorySurfaceRole.KNOWLEDGE_GRAPH,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.MEDIUM,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("concepts", "propositions", "prescriptions"),
+            canon_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            sqlite_count_tables=("concept_index", "propositions", "prescriptions"),
+        ),
+        SurfaceSpec(
+            "home.state_knowledge",
+            "{home}/.dharma/state/knowledge.db",
+            "dharma_swarm.knowledge_*",
+            MemorySurfaceRole.KNOWLEDGE_GRAPH,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("state_knowledge_atoms",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            do_not_migrate_reason="State knowledge DB is not canon without promotion.",
+        ),
+        SurfaceSpec(
+            "home.ontology",
+            "{home}/.dharma/ontology.db",
+            "dharma_swarm.ontology_hub",
+            MemorySurfaceRole.KNOWLEDGE_GRAPH,
+            SurfaceCategory.KNOWLEDGE_AUTHORITY,
+            AuthorityLevel.MEDIUM,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            provenance_quality="gated",
+            feeds=("ontology_objects", "lineage_edges"),
+            source_of_truth_for=("typed_ontology",),
+            canon_risk=RiskLevel.LOW,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            sqlite_count_tables=("objects", "links", "lineage_edges", "lineage_inputs", "lineage_outputs"),
+        ),
+        SurfaceSpec(
+            "home.temporal_graph",
+            "{home}/.dharma/db/temporal_graph.db",
+            "dharma_swarm.temporal_graph",
+            MemorySurfaceRole.PROJECTION_INDEX,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("temporal_mentions",),
+            projection_of=("memory_atoms", "concept_mentions"),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Temporal co-occurrence is evidence, not truth.",
+            sqlite_count_tables=("co_occurrences", "concept_sources", "concepts"),
+        ),
+        SurfaceSpec(
+            "home.dharma_graphs",
+            "{home}/.dharma/data/dharma_graphs.db",
+            "dharma_swarm.graph_store",
+            MemorySurfaceRole.KNOWLEDGE_GRAPH,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("semantic_graph_projection",),
+            projection_of=("runtime_events", "semantic_facts"),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Graph DB is projection/evidence unless gated elsewhere.",
+            sqlite_count_tables=("semantic_nodes", "semantic_edges"),
+        ),
+        SurfaceSpec(
+            "home.semantic_concept_graph",
+            "{home}/.dharma/semantic/concept_graph.json",
+            "dharma_swarm.vault_bridge",
+            MemorySurfaceRole.KNOWLEDGE_GRAPH,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("semantic_concept_graph_metadata",),
+            projection_of=("source_material", "memory_atoms"),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Semantic concept graph JSON is a projection, not authority.",
+        ),
+        SurfaceSpec(
+            "home.vectors",
+            "{home}/.dharma/vectors.db",
+            "dharma_swarm.vector_store",
+            MemorySurfaceRole.PROJECTION_INDEX,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.NONE,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("retrieval_pointers",),
+            projection_of=("source_chunks",),
+            canon_risk=RiskLevel.CRITICAL,
+            latency_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Vector stores are retrieval projections only.",
+            sqlite_count_tables=("vec_documents", "vec_embeddings_fallback", "vec_fts"),
+        ),
+        SurfaceSpec(
+            "home.lancedb",
+            "{home}/.dharma/lancedb/palace_docs.lance",
+            "dharma_swarm.memory_palace",
+            MemorySurfaceRole.PROJECTION_INDEX,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.NONE,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("retrieval_pointers",),
+            projection_of=("source_documents",),
+            canon_risk=RiskLevel.CRITICAL,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Huge LanceDB projection; expose metadata and pointers only.",
+        ),
+        SurfaceSpec(
+            "home.witness",
+            "{home}/.dharma/witness",
+            "dharma_swarm.witness",
+            MemorySurfaceRole.WITNESS,
+            SurfaceCategory.PROVENANCE,
+            AuthorityLevel.MEDIUM,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            feeds=("witness_events",),
+            source_of_truth_for=("audit_trails",),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.telos",
+            "{home}/.dharma/telos",
+            "dharma_swarm.telos_graph",
+            MemorySurfaceRole.INTENT_GRAPH,
+            SurfaceCategory.INTENT,
+            AuthorityLevel.MEDIUM,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            provenance_quality="gated",
+            feeds=("objectives", "strategies", "hypotheses", "key_results"),
+            source_of_truth_for=("approved_intent_graph",),
+            canon_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.agent_memory",
+            "{home}/.dharma/agent_memory",
+            "dharma_swarm.agent_memory",
+            MemorySurfaceRole.SESSION_MEMORY,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("agent_memory_atoms",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.LEGACY_TOLERATED,
+        ),
+        SurfaceSpec(
+            "home.agent_memory_db",
+            "{home}/.dharma/agent_memory/memories.db",
+            "dharma_swarm.agent_memory_manager",
+            MemorySurfaceRole.SESSION_MEMORY,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("agent_memory_atoms",),
+            canon_risk=RiskLevel.HIGH,
+            sqlite_count_tables=("memories", "shared_memories"),
+            migration_status=MigrationStatus.LEGACY_TOLERATED,
+        ),
+        SurfaceSpec(
+            "home.organism_memory",
+            "{home}/.dharma/organism_memory",
+            "dharma_swarm.organism_memory",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.STREAMING,
+            provenance_quality="weak",
+            known_writers=("organism_memory.py",),
+            feeds=("organism_self_model_events", "organism_self_model_edges"),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            do_not_migrate_reason="Organism self-model is evidence/reflection, not canon.",
+        ),
+        SurfaceSpec(
+            "home.messages",
+            "{home}/.dharma/db/messages.db",
+            "dharma_swarm.message_bus",
+            MemorySurfaceRole.COORDINATION,
+            SurfaceCategory.COORDINATION,
+            AuthorityLevel.MEDIUM,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("message_events", "artifacts"),
+            source_of_truth_for=("message_bus_state",),
+            sqlite_count_tables=("messages", "events", "subscriptions", "heartbeats", "artifacts"),
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.interop",
+            "{home}/.dharma/interop/message_bus.sqlite",
+            "dharma_swarm.operator_core.interop",
+            MemorySurfaceRole.COORDINATION,
+            SurfaceCategory.COORDINATION,
+            AuthorityLevel.MEDIUM,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("interop_tasks", "interop_events"),
+            sqlite_count_tables=("messages", "events", "heartbeats", "subscriptions", "artifacts"),
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.sessions",
+            "{home}/.dharma/sessions",
+            "dharma_swarm.operator_core.session_store",
+            MemorySurfaceRole.RAW_EPISODIC_LOG,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            feeds=("operator_sessions", "audit_events"),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.agent_interop",
+            "{home}/.dharma/agent_interop",
+            "dharma_swarm.operator_core.interop",
+            MemorySurfaceRole.COORDINATION,
+            SurfaceCategory.COORDINATION,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            feeds=("interop_events",),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.stigmergy",
+            "{home}/.dharma/stigmergy",
+            "dharma_swarm.stigmergy",
+            MemorySurfaceRole.COORDINATION,
+            SurfaceCategory.COORDINATION,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.STREAMING,
+            feeds=("coordination_marks",),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.traces",
+            "{home}/.dharma/traces",
+            "dharma_swarm.traces",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            feeds=("trace_events",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.evolution",
+            "{home}/.dharma/evolution",
+            "dharma_swarm.evolution",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            feeds=("evolution_memory", "fitness_learning"),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.population",
+            "{home}/.dharma/population",
+            "dharma_swarm.population_control",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("fitness_signals",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.training",
+            "{home}/.dharma/training",
+            "dharma_swarm.training_flywheel",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("training_trajectories",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.revenue_intel",
+            "{home}/.dharma/revenue_intel",
+            "dharma_swarm.orchestrate_live",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("revenue_intel_signals",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+    )
