+++ dharma_swarm/memory_kernel/surface_specs_extended.py
@@ -1,0 +1,605 @@
+"""Grouped MemoryKernel surface registry entries."""
+
+from __future__ import annotations
+
+from dharma_swarm.memory_kernel.atoms import (
+    AdapterMode,
+    AuthorityLevel,
+    MemorySurfaceRole,
+    MigrationStatus,
+    RiskLevel,
+    SurfaceCategory,
+    SurfaceStatus,
+    WriteMode,
+)
+from dharma_swarm.memory_kernel.surfaces import SurfaceSpec
+
+
+def extended_surface_specs() -> tuple[SurfaceSpec, ...]:
+    return (
+        SurfaceSpec(
+            "home.conversations",
+            "{home}/.dharma/conversations",
+            "api.routers.chat",
+            MemorySurfaceRole.RAW_EPISODIC_LOG,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            feeds=("dashboard_conversation_episodes",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.quality_gates",
+            "{home}/.dharma/quality_gates",
+            "dharma_swarm.quality_gates",
+            MemorySurfaceRole.WITNESS,
+            SurfaceCategory.PROVENANCE,
+            AuthorityLevel.MEDIUM,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.READ_ONLY,
+            feeds=("quality_gate_events",),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.evals",
+            "{home}/.dharma/evals",
+            "dharma_swarm.evaluator",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("evaluation_results",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.cost_log",
+            "{home}/.dharma/cost_log.jsonl",
+            "dharma_swarm.cost_tracker",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.STREAMING,
+            feeds=("cost_events",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.costs",
+            "{home}/.dharma/costs",
+            "dharma_swarm.cost_tracker",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.READ_ONLY,
+            feeds=("cost_reports",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.LOW,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.artifacts",
+            "{home}/.dharma/artifacts",
+            "dharma_swarm.artifact_store",
+            MemorySurfaceRole.RAW_EPISODIC_LOG,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("runtime_artifacts",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+            do_not_migrate_reason="Generated runtime artifacts are evidence, not audit truth.",
+        ),
+        SurfaceSpec(
+            "home.kaizen_ops",
+            "{home}/.dharma/kaizen/ops.db",
+            "dharma_swarm.integrations.kaizen_ops",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.MEDIUM,
+            WriteMode.READ_WRITE,
+            AdapterMode.READ_ONLY,
+            feeds=("ops_events",),
+            sqlite_count_tables=("events", "cron_health", "scout_health"),
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.ginko",
+            "{home}/.dharma/ginko",
+            "dharma_swarm.ginko_*",
+            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.STREAMING,
+            feeds=("market_memory", "ginko_logs"),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.meta",
+            "{home}/.dharma/meta",
+            "dharma_swarm.meta_*",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.MEDIUM,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.READ_ONLY,
+            feeds=("inventories", "gate_proposals", "concept_graphs"),
+            canon_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.meta_concept_graph",
+            "{home}/.dharma/meta/concept_graph.json",
+            "dharma_swarm.meta_*",
+            MemorySurfaceRole.KNOWLEDGE_GRAPH,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("meta_concept_graph_metadata",),
+            projection_of=("knowledge_atoms", "semantic_events"),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Meta concept graph is derived projection state.",
+        ),
+        SurfaceSpec(
+            "home.distilled",
+            "{home}/.dharma/distilled",
+            "dharma_swarm.conversation_distiller",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("distilled_conversation_metadata",),
+            projection_of=("conversation_logs",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Distilled outputs are summaries/projections until reviewed.",
+        ),
+        SurfaceSpec(
+            "home.shared",
+            "{home}/.dharma/shared",
+            "scripts/strange_loop.py",
+            MemorySurfaceRole.RAW_EPISODIC_LOG,
+            SurfaceCategory.RAW_EVIDENCE,
+            AuthorityLevel.LOW,
+            WriteMode.UNKNOWN,
+            AdapterMode.METADATA_ONLY,
+            feeds=("shared_memory_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.LEGACY_TOLERATED,
+            do_not_migrate_reason="Shared files are ad hoc memory outputs; classify before adapting.",
+        ),
+        SurfaceSpec(
+            "home.citations",
+            "{home}/.dharma/citations",
+            "dharma_swarm.citation_index",
+            MemorySurfaceRole.PROJECTION_INDEX,
+            SurfaceCategory.PROJECTION,
+            AuthorityLevel.LOW,
+            WriteMode.PROJECTION_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("citation_index_metadata",),
+            projection_of=("source_documents", "generated_reports"),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Citation indexes are projections over sources.",
+        ),
+        SurfaceSpec(
+            "home.subconscious",
+            "{home}/.dharma/subconscious",
+            "dharma_swarm.hypnagogic",
+            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("subconscious_dream_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.campaigns",
+            "{home}/.dharma/campaigns",
+            "dharma_swarm._campaign_manifest",
+            MemorySurfaceRole.INTENT_GRAPH,
+            SurfaceCategory.INTENT,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("campaign_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.workspace",
+            "{home}/.dharma/workspace",
+            "dharma_swarm.workspace_manager",
+            MemorySurfaceRole.RUNTIME_STATE,
+            SurfaceCategory.RUNTIME_CONTROL,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("workspace_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.active_inference",
+            "{home}/.dharma/active_inference",
+            "dharma_swarm.active_inference",
+            MemorySurfaceRole.RUNTIME_STATE,
+            SurfaceCategory.KNOWLEDGE_METABOLISM,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("active_inference_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.gaia_ledger",
+            "{home}/.dharma/gaia_ledger",
+            "dharma_swarm.gaia_ledger",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("gaia_ledger_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.gaia_platform",
+            "{home}/.dharma/gaia_platform",
+            "dharma_swarm.gaia_platform",
+            MemorySurfaceRole.RUNTIME_STATE,
+            SurfaceCategory.RUNTIME_CONTROL,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("gaia_platform_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.control",
+            "{home}/.dharma/control",
+            "scripts/consume_review_marks.py",
+            MemorySurfaceRole.COORDINATION,
+            SurfaceCategory.RUNTIME_CONTROL,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("control_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.cron",
+            "{home}/.dharma/cron",
+            "dharma_swarm.cron_*",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("cron_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.alerts",
+            "{home}/.dharma/alerts",
+            "dharma_swarm.monitor",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("alert_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.custodians",
+            "{home}/.dharma/custodians",
+            "dharma_swarm.custodians",
+            MemorySurfaceRole.COORDINATION,
+            SurfaceCategory.COORDINATION,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("custodian_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.jikoku",
+            "{home}/.dharma/jikoku",
+            "dharma_swarm.jikoku_*",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("jikoku_telemetry_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.terminal_supervisor",
+            "{home}/.dharma/terminal_supervisor",
+            "dharma_swarm.terminal_overnight_supervisor",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.READ_WRITE,
+            AdapterMode.METADATA_ONLY,
+            feeds=("terminal_supervisor_state_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.frontier_council",
+            "{home}/.dharma/frontier_council",
+            "external",
+            MemorySurfaceRole.SNAPSHOT,
+            SurfaceCategory.DORMANT,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.METADATA_ONLY,
+            active_status=SurfaceStatus.SNAPSHOT,
+            feeds=("snapshot_metadata",),
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Snapshot replicas; do not promote as live state.",
+        ),
+        SurfaceSpec(
+            "home.terminal_tui",
+            "{home}/.dharma/terminal_tui",
+            "dharma_swarm.terminal_bridge",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("terminal_telemetry_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Terminal TUI root can contain huge live logs; metadata only.",
+        ),
+        SurfaceSpec(
+            "home.terminal_tui_session_log",
+            "{home}/.dharma/terminal_tui/session.log",
+            "dharma_swarm.terminal_bridge",
+            MemorySurfaceRole.TELEMETRY,
+            SurfaceCategory.TELEMETRY,
+            AuthorityLevel.LOW,
+            WriteMode.APPEND_ONLY,
+            AdapterMode.METADATA_ONLY,
+            feeds=("terminal_telemetry",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Huge telemetry log; metadata only in M0/M1.",
+        ),
+        SurfaceSpec(
+            "home.claude_root",
+            "{home}/.claude",
+            "claude",
+            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.METADATA_ONLY,
+            feeds=("external_memory_metadata",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="External exocortex root; register children separately.",
+        ),
+        SurfaceSpec(
+            "home.claude_hooks",
+            "{home}/.claude/hooks",
+            "claude_hooks",
+            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            feeds=("hook_policy_context", "external_writers"),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.claude_projects",
+            "{home}/.claude/projects",
+            "claude",
+            MemorySurfaceRole.RAW_EPISODIC_LOG,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.STREAMING,
+            feeds=("external_episodes",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.HIGH,
+            latency_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.claude_cabinet",
+            "{home}/.claude/cabinet",
+            "claude",
+            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.MEDIUM,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            feeds=("external_docs",),
+            canon_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.claude_agent_memory",
+            "{home}/.claude/agent-memory",
+            "claude",
+            MemorySurfaceRole.SESSION_MEMORY,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            feeds=("external_agent_memory",),
+            canon_risk=RiskLevel.HIGH,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.codex_memory",
+            "{home}/.codex/memories/mcp-memory.jsonl",
+            "codex",
+            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            feeds=("external_memory_atoms",),
+            canon_risk=RiskLevel.HIGH,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.psmv",
+            "{home}/Persistent-Semantic-Memory-Vault",
+            "psmv",
+            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
+            SurfaceCategory.EXTERNAL,
+            AuthorityLevel.MEDIUM,
+            WriteMode.EXTERNAL,
+            AdapterMode.READ_ONLY,
+            feeds=("external_semantic_memory",),
+            canon_risk=RiskLevel.MEDIUM,
+            pii_secrets_risk=RiskLevel.MEDIUM,
+            latency_risk=RiskLevel.MEDIUM,
+            migration_status=MigrationStatus.ADAPTER_NEEDED,
+        ),
+        SurfaceSpec(
+            "home.m5_handoff",
+            "{home}/m5-handoff",
+            "m5_handoff",
+            MemorySurfaceRole.SNAPSHOT,
+            SurfaceCategory.DORMANT,
+            AuthorityLevel.LOW,
+            WriteMode.EXTERNAL,
+            AdapterMode.METADATA_ONLY,
+            active_status=SurfaceStatus.SNAPSHOT,
+            feeds=("handoff_snapshot_metadata",),
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Historical handoff bundle; compare, do not absorb.",
+        ),
+        SurfaceSpec(
+            "repo.dharma_root",
+            "{repo_root}/.dharma",
+            "repo_local",
+            MemorySurfaceRole.REPO_LOCAL_STATE,
+            SurfaceCategory.REPO_LOCAL,
+            AuthorityLevel.LOW,
+            WriteMode.UNKNOWN,
+            AdapterMode.METADATA_ONLY,
+            active_status=SurfaceStatus.SNAPSHOT,
+            feeds=("repo_state_metadata",),
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Repo-local state must not be confused with home state.",
+        ),
+        SurfaceSpec(
+            "repo.memory_plane",
+            "{repo_root}/.dharma/db/memory_plane.db",
+            "repo_local",
+            MemorySurfaceRole.REPO_LOCAL_STATE,
+            SurfaceCategory.REPO_LOCAL,
+            AuthorityLevel.LOW,
+            WriteMode.UNKNOWN,
+            AdapterMode.READ_ONLY,
+            active_status=SurfaceStatus.SNAPSHOT,
+            feeds=("repo_memory_plane_snapshot",),
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            sqlite_count_tables=("event_log", "conversation_turns", "idea_shards", "idea_links", "idea_uptake", "retrieval_log"),
+            do_not_migrate_reason="Repo-local snapshot; not live home memory.",
+        ),
+        SurfaceSpec(
+            "repo.runtime_state",
+            "{repo_root}/.dharma/state/runtime.db",
+            "repo_local",
+            MemorySurfaceRole.REPO_LOCAL_STATE,
+            SurfaceCategory.REPO_LOCAL,
+            AuthorityLevel.LOW,
+            WriteMode.UNKNOWN,
+            AdapterMode.READ_ONLY,
+            active_status=SurfaceStatus.SNAPSHOT,
+            feeds=("repo_runtime_snapshot",),
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            sqlite_count_tables=("session_events", "routing_decisions", "context_bundles", "memory_facts", "memory_edges"),
+            do_not_migrate_reason="Repo-local snapshot; not live home runtime.",
+        ),
+        SurfaceSpec(
+            "repo.swarm_memory",
+            "{repo_root}/.swarm/memory.db",
+            "repo_local",
+            MemorySurfaceRole.REPO_LOCAL_STATE,
+            SurfaceCategory.REPO_LOCAL,
+            AuthorityLevel.LOW,
+            WriteMode.UNKNOWN,
+            AdapterMode.READ_ONLY,
+            active_status=SurfaceStatus.SNAPSHOT,
+            feeds=("repo_swarm_snapshot",),
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            sqlite_count_tables=("memory_entries", "sessions", "patterns", "trajectory_steps"),
+            do_not_migrate_reason="Repo-local swarm memory is fixture/snapshot state.",
+        ),
+        SurfaceSpec(
+            "repo.literal_tilde_dharma",
+            "{repo_root}/~/.dharma",
+            "repo_local",
+            MemorySurfaceRole.UNSAFE_FIXTURE,
+            SurfaceCategory.UNSAFE,
+            AuthorityLevel.NONE,
+            WriteMode.UNKNOWN,
+            AdapterMode.METADATA_ONLY,
+            active_status=SurfaceStatus.UNSAFE,
+            feeds=("unsafe_fixture_metadata",),
+            canon_risk=RiskLevel.CRITICAL,
+            migration_status=MigrationStatus.DO_NOT_MIGRATE,
+            do_not_migrate_reason="Literal repo '~/.dharma' path is an unsafe fixture/snapshot, not the user's home.",
+        ),    )
