+++ dharma_swarm/memory_kernel/surfaces.py
@@ -43,8 +43,9 @@ class SurfaceSpec:
     metadata: dict[str, str] = field(default_factory=dict)
 
 
+
 def default_surface_specs() -> tuple[SurfaceSpec, ...]:
-    """Return the M0 known-surface registry.
+    """Return the known-surface registry.
 
     Templates support ``{home}`` and ``{repo_root}``.  The registry is explicit
     on purpose: projections, witness logs, runtime state, and external
@@ -52,1253 +53,7 @@ def default_surface_specs() -> tuple[SurfaceSpec, ...]:
     touch them.
     """
 
-    return (
-        SurfaceSpec(
-            "home.dharma_root",
-            "{home}/.dharma",
-            "external",
-            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.UNKNOWN,
-            AdapterMode.METADATA_ONLY,
-            feeds=("surface_health",),
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Root container; register children through adapters.",
-        ),
-        SurfaceSpec(
-            "home.memory_plane",
-            "{home}/.dharma/db/memory_plane.db",
-            "dharma_swarm.engine.event_memory",
-            MemorySurfaceRole.MEMORY_PLANE,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.MEDIUM,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.READ_ONLY,
-            provenance_quality="medium",
-            known_writers=(
-                "engine/event_memory.py",
-                "engine/conversation_memory.py",
-                "engine/retrieval_feedback.py",
-                "memory_lattice.py",
-            ),
-            known_readers=("context.py", "memory_lattice.py", "memory_palace.py"),
-            feeds=("episodes", "source_chunks", "retrieval_feedback", "idea_shards"),
-            source_of_truth_for=("raw_events", "conversation_turns", "idea_shards"),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            sqlite_count_tables=(
-                "event_log",
-                "conversation_turns",
-                "idea_shards",
-                "idea_links",
-                "idea_uptake",
-                "retrieval_log",
-                "source_documents",
-                "source_chunks",
-            ),
-        ),
-        SurfaceSpec(
-            "home.runtime_state",
-            "{home}/.dharma/state/runtime.db",
-            "dharma_swarm.runtime_state",
-            MemorySurfaceRole.RUNTIME_STATE,
-            SurfaceCategory.RUNTIME_CONTROL,
-            AuthorityLevel.HIGH,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            provenance_quality="medium",
-            known_writers=("runtime_state.py", "telemetry_plane.py", "route_witness.py"),
-            known_readers=("context_compiler.py", "terminal_bridge_context.py"),
-            feeds=("runtime_events", "memory_facts", "memory_edges", "context_bundles"),
-            source_of_truth_for=("runtime_sessions", "routing_decisions", "context_bundles"),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            sqlite_count_tables=(
-                "session_events",
-                "sessions",
-                "routing_decisions",
-                "context_bundles",
-                "memory_facts",
-                "memory_edges",
-            ),
-        ),
-        SurfaceSpec(
-            "home.memory_db",
-            "{home}/.dharma/db/memory.db",
-            "dharma_swarm.memory",
-            MemorySurfaceRole.SESSION_MEMORY,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            known_writers=("memory.py", "swarm.py", "pulse.py"),
-            known_readers=("context.py", "swarm.py", "mcp_server.py"),
-            feeds=("episodes",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.LEGACY_TOLERATED,
-            sqlite_count_tables=("memories",),
-        ),
-        SurfaceSpec(
-            "home.bridges",
-            "{home}/.dharma/db/bridges.db",
-            "dharma_swarm.bridge_registry",
-            MemorySurfaceRole.KNOWLEDGE_GRAPH,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            provenance_quality="medium",
-            known_writers=("bridge_registry.py", "pruner.py"),
-            feeds=("cross_graph_bridge_edges",),
-            projection_of=("semantic_graphs", "runtime_facts", "ontology", "telos"),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            do_not_migrate_reason="Bridge edges connect graphs but are not canon by themselves.",
-            sqlite_count_tables=("bridge_edges",),
-        ),
-        SurfaceSpec(
-            "home.ecosystem_index",
-            "{home}/.dharma/db/ecosystem_index.db",
-            "dharma_swarm.ecosystem_index",
-            MemorySurfaceRole.PROJECTION_INDEX,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.NONE,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.METADATA_ONLY,
-            known_writers=("ecosystem_index.py",),
-            feeds=("ecosystem_search_pointers",),
-            projection_of=("workspace_files", "external_docs"),
-            canon_risk=RiskLevel.CRITICAL,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Full-text index over files; source files remain authority.",
-            sqlite_count_tables=("doc_meta",),
-        ),
-        SurfaceSpec(
-            "home.router_audit_log",
-            "{home}/.dharma/logs/router/routing_decisions.jsonl",
-            "dharma_swarm.route_witness",
-            MemorySurfaceRole.WITNESS,
-            SurfaceCategory.PROVENANCE,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            provenance_quality="medium",
-            known_writers=("route_witness.py",),
-            known_readers=("routing dashboards", "post-hoc audits"),
-            feeds=("routing_witness_events", "provider_attempts"),
-            source_of_truth_for=("router_audit_jsonl",),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.routing_memory",
-            "{home}/.dharma/logs/router/routing_memory.sqlite3",
-            "dharma_swarm.routing_memory",
-            MemorySurfaceRole.RUNTIME_STATE,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            provenance_quality="medium",
-            known_writers=("routing_memory.py",),
-            known_readers=("smart_router.py", "agent_runner.py"),
-            feeds=("provider_learning", "routing_stats", "routing_events"),
-            source_of_truth_for=("provider_lane_learning",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            sqlite_count_tables=("routing_stats", "routing_events"),
-        ),
-        SurfaceSpec(
-            "home.hibernation",
-            "{home}/.dharma/state/hibernation.db",
-            "dharma_swarm.hibernation",
-            MemorySurfaceRole.RUNTIME_STATE,
-            SurfaceCategory.RUNTIME_CONTROL,
-            AuthorityLevel.MEDIUM,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            provenance_quality="medium",
-            known_writers=("hibernation.py",),
-            feeds=("hibernation_jobs", "wake_conditions"),
-            source_of_truth_for=("hibernateable_job_state",),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            sqlite_count_tables=("hibernation_jobs",),
-        ),
-        SurfaceSpec(
-            "home.algedonic_signals",
-            "{home}/.dharma/algedonic_signals.jsonl",
-            "dharma_swarm.orchestrator",
-            MemorySurfaceRole.WITNESS,
-            SurfaceCategory.PROVENANCE,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            provenance_quality="medium",
-            known_writers=("orchestrator.py",),
-            feeds=("algedonic_events",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.smriti",
-            "{home}/.smriti/smriti.db",
-            "smriti",
-            MemorySurfaceRole.SESSION_MEMORY,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            provenance_quality="weak",
-            known_writers=("~/.claude/hooks/smriti_session_end.py",),
-            known_readers=("~/.claude/hooks/smriti_session_start.py",),
-            feeds=("episodes", "procedural_candidates"),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            sqlite_count_tables=("memories", "entities", "relations"),
-        ),
-        SurfaceSpec(
-            "home.conversation_log",
-            "{home}/.dharma/conversation_log",
-            "dharma_swarm.conversation_log",
-            MemorySurfaceRole.RAW_EPISODIC_LOG,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            known_writers=("conversation_log.py", "~/.claude/hooks/conversation_logger.py"),
-            known_readers=("api/routers/dashboard_new.py",),
-            feeds=("episodes",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            do_not_migrate_reason="Use streaming adapter; never eager-load 24 GB logs.",
-        ),
-        SurfaceSpec(
-            "home.knowledge_root",
-            "{home}/.dharma/knowledge",
-            "chetana",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.METADATA_ONLY,
-            provenance_quality="medium",
-            feeds=("knowledge_lifecycle_atoms",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            do_not_migrate_reason="Chetana is metabolism; staging is not canon.",
-        ),
-        SurfaceSpec(
-            "home.knowledge_staging",
-            "{home}/.dharma/knowledge/staging",
-            "chetana",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.STREAMING,
-            feeds=("staged_knowledge_atoms",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            do_not_migrate_reason="Staging is capture backlog and must not be treated as canon.",
-        ),
-        SurfaceSpec(
-            "home.knowledge_wiki",
-            "{home}/.dharma/knowledge/wiki",
-            "chetana",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.MEDIUM,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            feeds=("curated_knowledge_atoms",),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.knowledge_quarantine",
-            "{home}/.dharma/knowledge/quarantine",
-            "chetana",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.NONE,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            feeds=("rejected_or_quarantined_atoms",),
-            canon_risk=RiskLevel.CRITICAL,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Quarantine is explicitly non-authoritative.",
-        ),
-        SurfaceSpec(
-            "home.knowledge_extracts",
-            "{home}/.dharma/knowledge/extracts",
-            "chetana",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.STREAMING,
-            feeds=("evidence_candidates",),
-            projection_of=("conversation_logs", "source_material"),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Extracts are derived evidence; stream pointers only.",
-        ),
-        SurfaceSpec(
-            "home.knowledge_store",
-            "{home}/.dharma/db/knowledge_store.db",
-            "dharma_swarm.engine.knowledge_store",
-            MemorySurfaceRole.KNOWLEDGE_GRAPH,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.MEDIUM,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("concepts", "propositions", "prescriptions"),
-            canon_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            sqlite_count_tables=("concept_index", "propositions", "prescriptions"),
-        ),
-        SurfaceSpec(
-            "home.state_knowledge",
-            "{home}/.dharma/state/knowledge.db",
-            "dharma_swarm.knowledge_*",
-            MemorySurfaceRole.KNOWLEDGE_GRAPH,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("state_knowledge_atoms",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            do_not_migrate_reason="State knowledge DB is not canon without promotion.",
-        ),
-        SurfaceSpec(
-            "home.ontology",
-            "{home}/.dharma/ontology.db",
-            "dharma_swarm.ontology_hub",
-            MemorySurfaceRole.KNOWLEDGE_GRAPH,
-            SurfaceCategory.KNOWLEDGE_AUTHORITY,
-            AuthorityLevel.MEDIUM,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            provenance_quality="gated",
-            feeds=("ontology_objects", "lineage_edges"),
-            source_of_truth_for=("typed_ontology",),
-            canon_risk=RiskLevel.LOW,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            sqlite_count_tables=("objects", "links", "lineage_edges", "lineage_inputs", "lineage_outputs"),
-        ),
-        SurfaceSpec(
-            "home.temporal_graph",
-            "{home}/.dharma/db/temporal_graph.db",
-            "dharma_swarm.temporal_graph",
-            MemorySurfaceRole.PROJECTION_INDEX,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("temporal_mentions",),
-            projection_of=("memory_atoms", "concept_mentions"),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Temporal co-occurrence is evidence, not truth.",
-            sqlite_count_tables=("co_occurrences", "concept_sources", "concepts"),
-        ),
-        SurfaceSpec(
-            "home.dharma_graphs",
-            "{home}/.dharma/data/dharma_graphs.db",
-            "dharma_swarm.graph_store",
-            MemorySurfaceRole.KNOWLEDGE_GRAPH,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("semantic_graph_projection",),
-            projection_of=("runtime_events", "semantic_facts"),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Graph DB is projection/evidence unless gated elsewhere.",
-            sqlite_count_tables=("semantic_nodes", "semantic_edges"),
-        ),
-        SurfaceSpec(
-            "home.semantic_concept_graph",
-            "{home}/.dharma/semantic/concept_graph.json",
-            "dharma_swarm.vault_bridge",
-            MemorySurfaceRole.KNOWLEDGE_GRAPH,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("semantic_concept_graph_metadata",),
-            projection_of=("source_material", "memory_atoms"),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Semantic concept graph JSON is a projection, not authority.",
-        ),
-        SurfaceSpec(
-            "home.vectors",
-            "{home}/.dharma/vectors.db",
-            "dharma_swarm.vector_store",
-            MemorySurfaceRole.PROJECTION_INDEX,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.NONE,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("retrieval_pointers",),
-            projection_of=("source_chunks",),
-            canon_risk=RiskLevel.CRITICAL,
-            latency_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Vector stores are retrieval projections only.",
-            sqlite_count_tables=("vec_documents", "vec_embeddings_fallback", "vec_fts"),
-        ),
-        SurfaceSpec(
-            "home.lancedb",
-            "{home}/.dharma/lancedb/palace_docs.lance",
-            "dharma_swarm.memory_palace",
-            MemorySurfaceRole.PROJECTION_INDEX,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.NONE,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("retrieval_pointers",),
-            projection_of=("source_documents",),
-            canon_risk=RiskLevel.CRITICAL,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Huge LanceDB projection; expose metadata and pointers only.",
-        ),
-        SurfaceSpec(
-            "home.witness",
-            "{home}/.dharma/witness",
-            "dharma_swarm.witness",
-            MemorySurfaceRole.WITNESS,
-            SurfaceCategory.PROVENANCE,
-            AuthorityLevel.MEDIUM,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            feeds=("witness_events",),
-            source_of_truth_for=("audit_trails",),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.telos",
-            "{home}/.dharma/telos",
-            "dharma_swarm.telos_graph",
-            MemorySurfaceRole.INTENT_GRAPH,
-            SurfaceCategory.INTENT,
-            AuthorityLevel.MEDIUM,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            provenance_quality="gated",
-            feeds=("objectives", "strategies", "hypotheses", "key_results"),
-            source_of_truth_for=("approved_intent_graph",),
-            canon_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.agent_memory",
-            "{home}/.dharma/agent_memory",
-            "dharma_swarm.agent_memory",
-            MemorySurfaceRole.SESSION_MEMORY,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("agent_memory_atoms",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.LEGACY_TOLERATED,
-        ),
-        SurfaceSpec(
-            "home.agent_memory_db",
-            "{home}/.dharma/agent_memory/memories.db",
-            "dharma_swarm.agent_memory_manager",
-            MemorySurfaceRole.SESSION_MEMORY,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("agent_memory_atoms",),
-            canon_risk=RiskLevel.HIGH,
-            sqlite_count_tables=("memories", "shared_memories"),
-            migration_status=MigrationStatus.LEGACY_TOLERATED,
-        ),
-        SurfaceSpec(
-            "home.organism_memory",
-            "{home}/.dharma/organism_memory",
-            "dharma_swarm.organism_memory",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.STREAMING,
-            provenance_quality="weak",
-            known_writers=("organism_memory.py",),
-            feeds=("organism_self_model_events", "organism_self_model_edges"),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            do_not_migrate_reason="Organism self-model is evidence/reflection, not canon.",
-        ),
-        SurfaceSpec(
-            "home.messages",
-            "{home}/.dharma/db/messages.db",
-            "dharma_swarm.message_bus",
-            MemorySurfaceRole.COORDINATION,
-            SurfaceCategory.COORDINATION,
-            AuthorityLevel.MEDIUM,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("message_events", "artifacts"),
-            source_of_truth_for=("message_bus_state",),
-            sqlite_count_tables=("messages", "events", "subscriptions", "heartbeats", "artifacts"),
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.interop",
-            "{home}/.dharma/interop/message_bus.sqlite",
-            "dharma_swarm.operator_core.interop",
-            MemorySurfaceRole.COORDINATION,
-            SurfaceCategory.COORDINATION,
-            AuthorityLevel.MEDIUM,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("interop_tasks", "interop_events"),
-            sqlite_count_tables=("messages", "events", "heartbeats", "subscriptions", "artifacts"),
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.sessions",
-            "{home}/.dharma/sessions",
-            "dharma_swarm.operator_core.session_store",
-            MemorySurfaceRole.RAW_EPISODIC_LOG,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            feeds=("operator_sessions", "audit_events"),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.agent_interop",
-            "{home}/.dharma/agent_interop",
-            "dharma_swarm.operator_core.interop",
-            MemorySurfaceRole.COORDINATION,
-            SurfaceCategory.COORDINATION,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            feeds=("interop_events",),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.stigmergy",
-            "{home}/.dharma/stigmergy",
-            "dharma_swarm.stigmergy",
-            MemorySurfaceRole.COORDINATION,
-            SurfaceCategory.COORDINATION,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.STREAMING,
-            feeds=("coordination_marks",),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.traces",
-            "{home}/.dharma/traces",
-            "dharma_swarm.traces",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            feeds=("trace_events",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.evolution",
-            "{home}/.dharma/evolution",
-            "dharma_swarm.evolution",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            feeds=("evolution_memory", "fitness_learning"),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.population",
-            "{home}/.dharma/population",
-            "dharma_swarm.population_control",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("fitness_signals",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.training",
-            "{home}/.dharma/training",
-            "dharma_swarm.training_flywheel",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("training_trajectories",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.revenue_intel",
-            "{home}/.dharma/revenue_intel",
-            "dharma_swarm.orchestrate_live",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("revenue_intel_signals",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.conversations",
-            "{home}/.dharma/conversations",
-            "api.routers.chat",
-            MemorySurfaceRole.RAW_EPISODIC_LOG,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            feeds=("dashboard_conversation_episodes",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.quality_gates",
-            "{home}/.dharma/quality_gates",
-            "dharma_swarm.quality_gates",
-            MemorySurfaceRole.WITNESS,
-            SurfaceCategory.PROVENANCE,
-            AuthorityLevel.MEDIUM,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.READ_ONLY,
-            feeds=("quality_gate_events",),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.evals",
-            "{home}/.dharma/evals",
-            "dharma_swarm.evaluator",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("evaluation_results",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.cost_log",
-            "{home}/.dharma/cost_log.jsonl",
-            "dharma_swarm.cost_tracker",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.STREAMING,
-            feeds=("cost_events",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.costs",
-            "{home}/.dharma/costs",
-            "dharma_swarm.cost_tracker",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.READ_ONLY,
-            feeds=("cost_reports",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.LOW,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.artifacts",
-            "{home}/.dharma/artifacts",
-            "dharma_swarm.artifact_store",
-            MemorySurfaceRole.RAW_EPISODIC_LOG,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("runtime_artifacts",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-            do_not_migrate_reason="Generated runtime artifacts are evidence, not audit truth.",
-        ),
-        SurfaceSpec(
-            "home.kaizen_ops",
-            "{home}/.dharma/kaizen/ops.db",
-            "dharma_swarm.integrations.kaizen_ops",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.MEDIUM,
-            WriteMode.READ_WRITE,
-            AdapterMode.READ_ONLY,
-            feeds=("ops_events",),
-            sqlite_count_tables=("events", "cron_health", "scout_health"),
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.ginko",
-            "{home}/.dharma/ginko",
-            "dharma_swarm.ginko_*",
-            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.STREAMING,
-            feeds=("market_memory", "ginko_logs"),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.meta",
-            "{home}/.dharma/meta",
-            "dharma_swarm.meta_*",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.MEDIUM,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.READ_ONLY,
-            feeds=("inventories", "gate_proposals", "concept_graphs"),
-            canon_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.meta_concept_graph",
-            "{home}/.dharma/meta/concept_graph.json",
-            "dharma_swarm.meta_*",
-            MemorySurfaceRole.KNOWLEDGE_GRAPH,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("meta_concept_graph_metadata",),
-            projection_of=("knowledge_atoms", "semantic_events"),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Meta concept graph is derived projection state.",
-        ),
-        SurfaceSpec(
-            "home.distilled",
-            "{home}/.dharma/distilled",
-            "dharma_swarm.conversation_distiller",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("distilled_conversation_metadata",),
-            projection_of=("conversation_logs",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Distilled outputs are summaries/projections until reviewed.",
-        ),
-        SurfaceSpec(
-            "home.shared",
-            "{home}/.dharma/shared",
-            "scripts/strange_loop.py",
-            MemorySurfaceRole.RAW_EPISODIC_LOG,
-            SurfaceCategory.RAW_EVIDENCE,
-            AuthorityLevel.LOW,
-            WriteMode.UNKNOWN,
-            AdapterMode.METADATA_ONLY,
-            feeds=("shared_memory_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.LEGACY_TOLERATED,
-            do_not_migrate_reason="Shared files are ad hoc memory outputs; classify before adapting.",
-        ),
-        SurfaceSpec(
-            "home.citations",
-            "{home}/.dharma/citations",
-            "dharma_swarm.citation_index",
-            MemorySurfaceRole.PROJECTION_INDEX,
-            SurfaceCategory.PROJECTION,
-            AuthorityLevel.LOW,
-            WriteMode.PROJECTION_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("citation_index_metadata",),
-            projection_of=("source_documents", "generated_reports"),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Citation indexes are projections over sources.",
-        ),
-        SurfaceSpec(
-            "home.subconscious",
-            "{home}/.dharma/subconscious",
-            "dharma_swarm.hypnagogic",
-            MemorySurfaceRole.KNOWLEDGE_LIFECYCLE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("subconscious_dream_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.campaigns",
-            "{home}/.dharma/campaigns",
-            "dharma_swarm._campaign_manifest",
-            MemorySurfaceRole.INTENT_GRAPH,
-            SurfaceCategory.INTENT,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("campaign_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.workspace",
-            "{home}/.dharma/workspace",
-            "dharma_swarm.workspace_manager",
-            MemorySurfaceRole.RUNTIME_STATE,
-            SurfaceCategory.RUNTIME_CONTROL,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("workspace_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.active_inference",
-            "{home}/.dharma/active_inference",
-            "dharma_swarm.active_inference",
-            MemorySurfaceRole.RUNTIME_STATE,
-            SurfaceCategory.KNOWLEDGE_METABOLISM,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("active_inference_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.gaia_ledger",
-            "{home}/.dharma/gaia_ledger",
-            "dharma_swarm.gaia_ledger",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("gaia_ledger_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.gaia_platform",
-            "{home}/.dharma/gaia_platform",
-            "dharma_swarm.gaia_platform",
-            MemorySurfaceRole.RUNTIME_STATE,
-            SurfaceCategory.RUNTIME_CONTROL,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("gaia_platform_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.control",
-            "{home}/.dharma/control",
-            "scripts/consume_review_marks.py",
-            MemorySurfaceRole.COORDINATION,
-            SurfaceCategory.RUNTIME_CONTROL,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("control_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.cron",
-            "{home}/.dharma/cron",
-            "dharma_swarm.cron_*",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("cron_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.alerts",
-            "{home}/.dharma/alerts",
-            "dharma_swarm.monitor",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("alert_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.custodians",
-            "{home}/.dharma/custodians",
-            "dharma_swarm.custodians",
-            MemorySurfaceRole.COORDINATION,
-            SurfaceCategory.COORDINATION,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("custodian_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.jikoku",
-            "{home}/.dharma/jikoku",
-            "dharma_swarm.jikoku_*",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("jikoku_telemetry_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.terminal_supervisor",
-            "{home}/.dharma/terminal_supervisor",
-            "dharma_swarm.terminal_overnight_supervisor",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.READ_WRITE,
-            AdapterMode.METADATA_ONLY,
-            feeds=("terminal_supervisor_state_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.frontier_council",
-            "{home}/.dharma/frontier_council",
-            "external",
-            MemorySurfaceRole.SNAPSHOT,
-            SurfaceCategory.DORMANT,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.METADATA_ONLY,
-            active_status=SurfaceStatus.SNAPSHOT,
-            feeds=("snapshot_metadata",),
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Snapshot replicas; do not promote as live state.",
-        ),
-        SurfaceSpec(
-            "home.terminal_tui",
-            "{home}/.dharma/terminal_tui",
-            "dharma_swarm.terminal_bridge",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("terminal_telemetry_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Terminal TUI root can contain huge live logs; metadata only.",
-        ),
-        SurfaceSpec(
-            "home.terminal_tui_session_log",
-            "{home}/.dharma/terminal_tui/session.log",
-            "dharma_swarm.terminal_bridge",
-            MemorySurfaceRole.TELEMETRY,
-            SurfaceCategory.TELEMETRY,
-            AuthorityLevel.LOW,
-            WriteMode.APPEND_ONLY,
-            AdapterMode.METADATA_ONLY,
-            feeds=("terminal_telemetry",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Huge telemetry log; metadata only in M0/M1.",
-        ),
-        SurfaceSpec(
-            "home.claude_root",
-            "{home}/.claude",
-            "claude",
-            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.METADATA_ONLY,
-            feeds=("external_memory_metadata",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="External exocortex root; register children separately.",
-        ),
-        SurfaceSpec(
-            "home.claude_hooks",
-            "{home}/.claude/hooks",
-            "claude_hooks",
-            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            feeds=("hook_policy_context", "external_writers"),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.claude_projects",
-            "{home}/.claude/projects",
-            "claude",
-            MemorySurfaceRole.RAW_EPISODIC_LOG,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.STREAMING,
-            feeds=("external_episodes",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.HIGH,
-            latency_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.claude_cabinet",
-            "{home}/.claude/cabinet",
-            "claude",
-            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.MEDIUM,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            feeds=("external_docs",),
-            canon_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.claude_agent_memory",
-            "{home}/.claude/agent-memory",
-            "claude",
-            MemorySurfaceRole.SESSION_MEMORY,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            feeds=("external_agent_memory",),
-            canon_risk=RiskLevel.HIGH,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.codex_memory",
-            "{home}/.codex/memories/mcp-memory.jsonl",
-            "codex",
-            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            feeds=("external_memory_atoms",),
-            canon_risk=RiskLevel.HIGH,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.psmv",
-            "{home}/Persistent-Semantic-Memory-Vault",
-            "psmv",
-            MemorySurfaceRole.EXTERNAL_EXOCORTEX,
-            SurfaceCategory.EXTERNAL,
-            AuthorityLevel.MEDIUM,
-            WriteMode.EXTERNAL,
-            AdapterMode.READ_ONLY,
-            feeds=("external_semantic_memory",),
-            canon_risk=RiskLevel.MEDIUM,
-            pii_secrets_risk=RiskLevel.MEDIUM,
-            latency_risk=RiskLevel.MEDIUM,
-            migration_status=MigrationStatus.ADAPTER_NEEDED,
-        ),
-        SurfaceSpec(
-            "home.m5_handoff",
-            "{home}/m5-handoff",
-            "m5_handoff",
-            MemorySurfaceRole.SNAPSHOT,
-            SurfaceCategory.DORMANT,
-            AuthorityLevel.LOW,
-            WriteMode.EXTERNAL,
-            AdapterMode.METADATA_ONLY,
-            active_status=SurfaceStatus.SNAPSHOT,
-            feeds=("handoff_snapshot_metadata",),
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Historical handoff bundle; compare, do not absorb.",
-        ),
-        SurfaceSpec(
-            "repo.dharma_root",
-            "{repo_root}/.dharma",
-            "repo_local",
-            MemorySurfaceRole.REPO_LOCAL_STATE,
-            SurfaceCategory.REPO_LOCAL,
-            AuthorityLevel.LOW,
-            WriteMode.UNKNOWN,
-            AdapterMode.METADATA_ONLY,
-            active_status=SurfaceStatus.SNAPSHOT,
-            feeds=("repo_state_metadata",),
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Repo-local state must not be confused with home state.",
-        ),
-        SurfaceSpec(
-            "repo.memory_plane",
-            "{repo_root}/.dharma/db/memory_plane.db",
-            "repo_local",
-            MemorySurfaceRole.REPO_LOCAL_STATE,
-            SurfaceCategory.REPO_LOCAL,
-            AuthorityLevel.LOW,
-            WriteMode.UNKNOWN,
-            AdapterMode.READ_ONLY,
-            active_status=SurfaceStatus.SNAPSHOT,
-            feeds=("repo_memory_plane_snapshot",),
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            sqlite_count_tables=("event_log", "conversation_turns", "idea_shards", "idea_links", "idea_uptake", "retrieval_log"),
-            do_not_migrate_reason="Repo-local snapshot; not live home memory.",
-        ),
-        SurfaceSpec(
-            "repo.runtime_state",
-            "{repo_root}/.dharma/state/runtime.db",
-            "repo_local",
-            MemorySurfaceRole.REPO_LOCAL_STATE,
-            SurfaceCategory.REPO_LOCAL,
-            AuthorityLevel.LOW,
-            WriteMode.UNKNOWN,
-            AdapterMode.READ_ONLY,
-            active_status=SurfaceStatus.SNAPSHOT,
-            feeds=("repo_runtime_snapshot",),
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            sqlite_count_tables=("session_events", "routing_decisions", "context_bundles", "memory_facts", "memory_edges"),
-            do_not_migrate_reason="Repo-local snapshot; not live home runtime.",
-        ),
-        SurfaceSpec(
-            "repo.swarm_memory",
-            "{repo_root}/.swarm/memory.db",
-            "repo_local",
-            MemorySurfaceRole.REPO_LOCAL_STATE,
-            SurfaceCategory.REPO_LOCAL,
-            AuthorityLevel.LOW,
-            WriteMode.UNKNOWN,
-            AdapterMode.READ_ONLY,
-            active_status=SurfaceStatus.SNAPSHOT,
-            feeds=("repo_swarm_snapshot",),
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            sqlite_count_tables=("memory_entries", "sessions", "patterns", "trajectory_steps"),
-            do_not_migrate_reason="Repo-local swarm memory is fixture/snapshot state.",
-        ),
-        SurfaceSpec(
-            "repo.literal_tilde_dharma",
-            "{repo_root}/~/.dharma",
-            "repo_local",
-            MemorySurfaceRole.UNSAFE_FIXTURE,
-            SurfaceCategory.UNSAFE,
-            AuthorityLevel.NONE,
-            WriteMode.UNKNOWN,
-            AdapterMode.METADATA_ONLY,
-            active_status=SurfaceStatus.UNSAFE,
-            feeds=("unsafe_fixture_metadata",),
-            canon_risk=RiskLevel.CRITICAL,
-            migration_status=MigrationStatus.DO_NOT_MIGRATE,
-            do_not_migrate_reason="Literal repo '~/.dharma' path is an unsafe fixture/snapshot, not the user's home.",
-        ),
-    )
+    from dharma_swarm.memory_kernel.surface_specs_core import core_surface_specs
+    from dharma_swarm.memory_kernel.surface_specs_extended import extended_surface_specs
+
+    return (*core_surface_specs(), *extended_surface_specs())
