+++ dharma_swarm/memory_kernel/writer_specs.py
@@ -1,0 +1,589 @@
+"""Registered memory writer inventory for MemoryKernel."""
+
+from __future__ import annotations
+
+from dharma_swarm.memory_kernel.atoms import RiskLevel, WriteMode
+from dharma_swarm.memory_kernel.writer_models import (
+    MemoryWriterSpec,
+    WriterClassification,
+)
+
+
+def default_writer_specs() -> tuple[MemoryWriterSpec, ...]:
+    return (
+        MemoryWriterSpec(
+            "memory_lattice.ingest_runtime_envelope",
+            "dharma_swarm.memory_lattice",
+            "MemoryLattice.ingest_runtime_envelope",
+            ("home.runtime_state",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Coordinates event ingestion and optional JSONL event log append.",
+        ),
+        MemoryWriterSpec(
+            "memory_lattice.remember",
+            "dharma_swarm.memory_lattice",
+            "MemoryLattice.remember",
+            ("home.runtime_state", "home.vectors"),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Writes strange-loop memory and projection index entries.",
+        ),
+        MemoryWriterSpec(
+            "memory_lattice.record_fact",
+            "dharma_swarm.memory_lattice",
+            "MemoryLattice.record_fact",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Fact writes must remain non-canonical until KnowledgeOps promotion gates exist.",
+        ),
+        MemoryWriterSpec(
+            "runtime_state.record_memory_fact",
+            "dharma_swarm.runtime_state",
+            "RuntimeStateStore.record_memory_fact",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.APPROVED,
+            RiskLevel.MEDIUM,
+            "Control-plane fact store; authority is runtime, not semantic canon.",
+        ),
+        MemoryWriterSpec(
+            "runtime_state.record_memory_edge",
+            "dharma_swarm.runtime_state",
+            "RuntimeStateStore.record_memory_edge",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.APPROVED,
+            RiskLevel.MEDIUM,
+            "Control-plane edge store; authority is runtime, not semantic canon.",
+        ),
+        MemoryWriterSpec(
+            "conversation_memory.record_turn",
+            "dharma_swarm.engine.conversation_memory",
+            "ConversationMemoryStore.record_turn",
+            ("home.memory_plane",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.HIGH,
+            "Raw conversation and shard harvesting path; keep provenance labels attached.",
+        ),
+        MemoryWriterSpec(
+            "retrieval_feedback.log_hits",
+            "dharma_swarm.engine.retrieval_feedback",
+            "RetrievalFeedbackStore.log_hits",
+            ("home.memory_plane",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.MEDIUM,
+            "Retrieval telemetry is feedback evidence, not truth.",
+        ),
+        MemoryWriterSpec(
+            "providers.routing_audit",
+            "dharma_swarm.providers",
+            "ModelRouter._append_routing_audit",
+            ("home.router_audit_log",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Provider router audit JSONL is route evidence, not semantic truth.",
+        ),
+        MemoryWriterSpec(
+            "providers.route_retrospective",
+            "dharma_swarm.providers",
+            "ModelRouter._append_route_retrospective",
+            ("home.router_audit_log",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Route retrospective JSONL is provider-learning evidence, not canon.",
+        ),
+        MemoryWriterSpec(
+            "routing_memory.store",
+            "dharma_swarm.routing_memory",
+            "RoutingMemoryStore",
+            ("home.routing_memory",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Provider-learning SQLite store is memory-like runtime telemetry, not semantic truth.",
+        ),
+        MemoryWriterSpec(
+            "runtime_state.store",
+            "dharma_swarm.runtime_state",
+            "RuntimeStateStore",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.APPROVED,
+            RiskLevel.MEDIUM,
+            "RuntimeStateStore is the runtime/control-plane writer boundary, not semantic canon.",
+        ),
+        MemoryWriterSpec(
+            "telemetry_plane.store",
+            "dharma_swarm.telemetry_plane",
+            "TelemetryPlaneStore",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Telemetry plane writes company-state records into runtime.db.",
+        ),
+        MemoryWriterSpec(
+            "event_memory.store",
+            "dharma_swarm.engine.event_memory",
+            "EventMemoryStore",
+            ("home.memory_plane",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.MEDIUM,
+            "Event memory is raw runtime evidence in memory_plane.db.",
+        ),
+        MemoryWriterSpec(
+            "conversation_memory.store",
+            "dharma_swarm.engine.conversation_memory",
+            "ConversationMemoryStore",
+            ("home.memory_plane",),
+            WriteMode.READ_WRITE,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.HIGH,
+            "Conversation turns, idea shards, uptake, and follow-ups are raw evidence.",
+        ),
+        MemoryWriterSpec(
+            "retrieval_feedback.store",
+            "dharma_swarm.engine.retrieval_feedback",
+            "RetrievalFeedbackStore",
+            ("home.memory_plane",),
+            WriteMode.READ_WRITE,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.MEDIUM,
+            "Retrieval feedback is usefulness evidence, not truth.",
+        ),
+        MemoryWriterSpec(
+            "unified_index.store",
+            "dharma_swarm.engine.unified_index",
+            "UnifiedIndex",
+            ("home.memory_plane",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Unified index writes documents/chunks into memory_plane as retrieval projection state.",
+        ),
+        MemoryWriterSpec(
+            "vector_store.store",
+            "dharma_swarm.vector_store",
+            "VectorStore",
+            ("home.vectors",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "VectorStore writes retrieval projections; never authority.",
+        ),
+        MemoryWriterSpec(
+            "vector_store.tfidf_state",
+            "dharma_swarm.vector_store",
+            "TFIDFEmbedder",
+            ("home.vectors",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "TF-IDF model state is retrieval projection sidecar state.",
+        ),
+        MemoryWriterSpec(
+            "agent_memory_manager.store",
+            "dharma_swarm.agent_memory_manager",
+            "AgentMemoryManager",
+            ("home.agent_memory_db",),
+            WriteMode.READ_WRITE,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.HIGH,
+            "SQLite-backed agent memories remain raw agent evidence.",
+        ),
+        MemoryWriterSpec(
+            "organism_memory.store",
+            "dharma_swarm.organism_memory",
+            "OrganismMemory",
+            ("home.organism_memory",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Organism self-model events and relationships are reflections, not canon.",
+        ),
+        MemoryWriterSpec(
+            "swarm.manager_memory_writers",
+            "dharma_swarm.swarm",
+            "SwarmManager",
+            (
+                "home.memory_db",
+                "home.memory_plane",
+                "home.runtime_state",
+                "home.agent_memory",
+            ),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Top-level coordinator delegates to multiple memory stores; keep boundary visible.",
+        ),
+        MemoryWriterSpec(
+            "intelligence_adapters.memory_plane",
+            "dharma_swarm.contracts.intelligence_adapters",
+            "SovereignMemoryPlaneAdapter",
+            ("home.memory_plane",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Contract adapter writes to the memory plane; provenance must stay attached.",
+        ),
+        MemoryWriterSpec(
+            "intelligence_adapters.skill_store",
+            "dharma_swarm.contracts.intelligence_adapters",
+            "SovereignSkillStoreAdapter",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Skill-store contract state is runtime/control evidence.",
+        ),
+        MemoryWriterSpec(
+            "intelligence_adapters.evaluation_sink",
+            "dharma_swarm.contracts.intelligence_adapters",
+            "SovereignEvaluationSinkAdapter",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Evaluation sink writes runtime evaluation telemetry.",
+        ),
+        MemoryWriterSpec(
+            "intelligence_adapters.learning_engine",
+            "dharma_swarm.contracts.intelligence_adapters",
+            "SovereignLearningEngineAdapter",
+            ("home.runtime_state",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Learning-engine contract writes runtime learning records.",
+        ),
+        MemoryWriterSpec(
+            "bridge_registry.store",
+            "dharma_swarm.bridge_registry",
+            "BridgeRegistry",
+            ("home.bridges",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Cross-graph bridge edges are linking evidence, not canon.",
+        ),
+        MemoryWriterSpec(
+            "ecosystem_index.store",
+            "dharma_swarm.ecosystem_index",
+            "EcosystemIndex",
+            ("home.ecosystem_index",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "FTS ecosystem index is a projection over source files.",
+        ),
+        MemoryWriterSpec(
+            "graph_store.sqlite",
+            "dharma_swarm.graph_store",
+            "SQLiteGraphStore",
+            ("home.dharma_graphs",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Generic graph-store writes derived graph state.",
+        ),
+        MemoryWriterSpec(
+            "hibernation.store",
+            "dharma_swarm.hibernation",
+            "HibernationManager",
+            ("home.hibernation",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Hibernate-and-wake state is runtime control memory.",
+        ),
+        MemoryWriterSpec(
+            "kaizen_ops_local.store",
+            "dharma_swarm.kaizen_ops_local",
+            "KaizenOpsLocal",
+            ("home.kaizen_ops",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Kaizen ops database is operational telemetry.",
+        ),
+        MemoryWriterSpec(
+            "knowledge_units.store",
+            "dharma_swarm.knowledge_units",
+            "KnowledgeStore",
+            ("home.state_knowledge",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Structured knowledge units are candidates until KnowledgeOps promotion gates exist.",
+        ),
+        MemoryWriterSpec(
+            "lineage.graph",
+            "dharma_swarm.lineage",
+            "LineageGraph",
+            ("home.ontology",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Lineage graph writes provenance edges into the ontology runtime store.",
+        ),
+        MemoryWriterSpec(
+            "ontology_hub.store",
+            "dharma_swarm.ontology_hub",
+            "OntologyHub",
+            ("home.ontology",),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Ontology hub is semantic graph state but not M0 canonical by default.",
+        ),
+        MemoryWriterSpec(
+            "temporal_graph.store",
+            "dharma_swarm.temporal_graph",
+            "TemporalKnowledgeGraph",
+            ("home.temporal_graph",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Temporal co-occurrence graph is temporal evidence/projection.",
+        ),
+        MemoryWriterSpec(
+            "pruner.store",
+            "dharma_swarm.pruner",
+            "Pruner",
+            ("home.stigmergy", "home.bridges"),
+            WriteMode.READ_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Pruner mutates coordination marks and bridge edges; retain audit visibility.",
+        ),
+        MemoryWriterSpec(
+            "observability.local_trace_store",
+            "dharma_swarm.observability",
+            "LocalTraceStore",
+            ("home.traces",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Local traces and cost ledger entries are telemetry evidence.",
+        ),
+        MemoryWriterSpec(
+            "agent_memory.remember",
+            "dharma_swarm.agent_memory",
+            "AgentMemoryBank.remember",
+            ("home.agent_memory",),
+            WriteMode.READ_WRITE,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.HIGH,
+            "Agent self-memory remains raw evidence until gated.",
+        ),
+        MemoryWriterSpec(
+            "agent_memory.save",
+            "dharma_swarm.agent_memory",
+            "AgentMemoryBank.save",
+            ("home.agent_memory",),
+            WriteMode.READ_WRITE,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.HIGH,
+            "JSON persistence for agent self-memory.",
+        ),
+        MemoryWriterSpec(
+            "conversation_log.log_exchange",
+            "dharma_swarm.conversation_log",
+            "log_exchange",
+            ("home.conversation_log",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.HIGH,
+            "High-volume raw episodic log; must stay streaming/metadata-first.",
+        ),
+        MemoryWriterSpec(
+            "conversation_log.log_agent_turn",
+            "dharma_swarm.conversation_log",
+            "log_agent_turn",
+            ("home.conversation_log",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.LEGACY_TOLERATED,
+            RiskLevel.HIGH,
+            "High-volume raw episodic log; must stay streaming/metadata-first.",
+        ),
+        MemoryWriterSpec(
+            "semantic_memory_bridge.index_concepts",
+            "dharma_swarm.semantic_memory_bridge",
+            "index_concepts_into_memory",
+            ("home.runtime_state",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Semantic concepts are indexed into retrieval; they are derived, not canon.",
+        ),
+        MemoryWriterSpec(
+            "semantic_memory_bridge.sleep_phase",
+            "dharma_swarm.semantic_memory_bridge",
+            "run_semantic_sleep_phase",
+            ("home.semantic_concept_graph",),
+            WriteMode.PROJECTION_WRITE,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.HIGH,
+            "Semantic sleep graph is a projection and should not be cited as authority.",
+        ),
+        MemoryWriterSpec(
+            "algedonic_bridge.write_witness",
+            "dharma_swarm.algedonic_bridge",
+            "_write_witness",
+            ("home.witness",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.DORMANT,
+            RiskLevel.MEDIUM,
+            "Algedonic bridge writer is planned or branch-local on current main.",
+        ),
+        MemoryWriterSpec(
+            "bhed_gnan_monitor.write_witness",
+            "dharma_swarm.bhed_gnan_monitor",
+            "BhedGnanMonitor.write_witness",
+            ("home.witness",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.DORMANT,
+            RiskLevel.MEDIUM,
+            "Bhed/Gnani monitor writer is planned or branch-local on current main.",
+        ),
+        MemoryWriterSpec(
+            "persistent_agent.write_witness",
+            "dharma_swarm.persistent_agent",
+            "PersistentAgent._write_witness",
+            ("home.witness",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Persistent-agent witness append path.",
+        ),
+        MemoryWriterSpec(
+            "replication_protocol.write_witness",
+            "dharma_swarm.replication_protocol",
+            "ReplicationProtocol._write_witness",
+            ("home.witness",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Replication witness append path.",
+        ),
+        MemoryWriterSpec(
+            "telos_gates.write_witness",
+            "dharma_swarm.telos_gates",
+            "TelosGatekeeper._log_witness",
+            ("home.telos", "home.witness"),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Telos witness logs are intent/provenance evidence.",
+        ),
+        MemoryWriterSpec(
+            "telos_gates.coupling_feedback",
+            "dharma_swarm.telos_gates",
+            "check_action",
+            ("home.meta",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Gate feedback coupling is meta-control evidence.",
+        ),
+        MemoryWriterSpec(
+            "opportunity_dispatcher.review_warn",
+            "dharma_swarm.opportunity_dispatcher",
+            "_write_review_warn",
+            ("home.witness",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.DORMANT,
+            RiskLevel.MEDIUM,
+            "Opportunity dispatcher writer is planned or branch-local on current main.",
+        ),
+        MemoryWriterSpec(
+            "orchestrator.task_failure_algedonic",
+            "dharma_swarm.orchestrator",
+            "Orchestrator._handle_task_failure",
+            ("home.algedonic_signals",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Task failure algedonic signal append path.",
+        ),
+        MemoryWriterSpec(
+            "consume_review_marks.append_witness",
+            "scripts.consume_review_marks",
+            "append_witness",
+            ("home.control",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.DORMANT,
+            RiskLevel.MEDIUM,
+            "Review mark consumer script is not present on current main.",
+        ),
+        MemoryWriterSpec(
+            "consume_review_marks.escalate_to_taskboard",
+            "scripts.consume_review_marks",
+            "escalate_to_taskboard",
+            ("home.control",),
+            WriteMode.READ_WRITE,
+            WriterClassification.DORMANT,
+            RiskLevel.MEDIUM,
+            "Review mark consumer script is not present on current main.",
+        ),
+        MemoryWriterSpec(
+            "check_packet_provenance.append_witness",
+            "scripts.governance.check_packet_provenance",
+            "append_witness",
+            ("home.witness",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.DORMANT,
+            RiskLevel.MEDIUM,
+            "Packet provenance checker script is not present on current main.",
+        ),
+        MemoryWriterSpec(
+            "orchestrate_live.revenue_intel_signal",
+            "dharma_swarm.orchestrate_live",
+            "_on_revenue_intel",
+            ("home.revenue_intel",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Signal-bus revenue intel JSONL is telemetry evidence, not canon.",
+        ),
+        MemoryWriterSpec(
+            "population_control.record_fitness_signal",
+            "dharma_swarm.population_control",
+            "record_fitness_signal",
+            ("home.population",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Fitness signal JSONL is operational learning evidence.",
+        ),
+        MemoryWriterSpec(
+            "training_flywheel.record_trajectory_from_signal",
+            "dharma_swarm.training_flywheel",
+            "record_trajectory_from_signal",
+            ("home.training",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Training trajectory JSONL is learning telemetry.",
+        ),
+        MemoryWriterSpec(
+            "witness.record_anomaly_signal",
+            "dharma_swarm.witness",
+            "record_anomaly_signal",
+            ("home.witness",),
+            WriteMode.APPEND_ONLY,
+            WriterClassification.REVIEW_REQUIRED,
+            RiskLevel.MEDIUM,
+            "Anomaly signal JSONL is witness evidence.",
+        ),
+    )
